# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KeyStar Platform API
  description: |
    Public client API and server-to-server API of the KeyStar licensing platform.

    Every request resolves to an application through the `X-KeyStar-App`
    header plus a bearer credential. Two credential types exist:

    * **Publishable key** (`ks_pk_test_...` / `ks_pk_live_...`) — ships inside
      desktop or mobile clients; authorises login, device verification and
      public configuration only.
    * **Secret key** (`ks_sk_test_...` / `ks_sk_sk_...` style prefix
      `ks_sk_...`) — server-side only; scoped to explicit permissions such as
      `users.read`, `licenses.write`.

    Never embed a secret key in a distributed application.
  version: 1.0.0
servers:
  - url: https://your-keystar-host
    description: Self-hosted deployment
...
security:
  - AppHeader: []
  - PublishableKey: []
  - SecretKey: []
paths:
  /v1/auth/login:
    post:
      tags: [Client Auth]
      summary: Authenticate an end user
      description: |
        Authenticates with email/password and returns a pending device
        challenge when the device is unknown. Successful completion yields a
        short-lived access token plus refresh token. Rate limited per IP and
        credential.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                device_fingerprint:
                  type: string
                  description: Aggregate hardware fingerprint produced by the SDK.
      responses:
        "200":
          description: Login accepted. Returns tokens directly for trusted devices, otherwise a pending challenge.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/device/verify:
    post:
      tags: [Client Auth]
      summary: Complete device verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                challenge_id: { type: string }
                fingerprint: { type: string }
      responses:
        "200":
          description: Device verified; session tokens returned.
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/auth/refresh:
    post:
      tags: [Client Auth]
      summary: Rotate a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200": { description: New token pair issued. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/me:
    get:
      tags: [Client Auth]
      summary: Profile of the authenticated end user
      security:
        - AppHeader: []
      responses:
        "200": { description: Current user profile. }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    AppHeader:
      type: apiKey
      in: header
      name: X-KeyStar-App
      description: Application UUID that forms the tenant boundary.
    PublishableKey:
      type: http
      scheme: bearer
      description: "`ks_pk_*` client credential."
    SecretKey:
      type: http
      scheme: bearer
      description: "`ks_sk_*` server credential with explicit scopes."
  responses:
    Unauthorized:
      description: Missing or invalid credential.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Per-credential or per-IP rate limit exceeded. Retry using the Retry-After header.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds until the request may be retried.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        code: { type: string, example: INVALID_CREDENTIAL }
        message: { type: string }
        request_id: { type: string, format: uuid }
    LoginResult:
      type: object
      description: Either a completed token pair or a pending device challenge.
      properties:
        ok: { type: boolean }
        status:
          type: string
          enum: [authenticated, device_challenge]
        access_token: { type: string }
        refresh_token: { type: string }
        expires_at: { type: string, format: date-time }
        challenge:
          type: object
          nullable: true